scripts/cartesia-line-agent/pyproject.toml
[project]
name = "guided-growth-line-agent"
version = "0.1.0"
requires-python = ">=3.9"
dependencies = [
  "cartesia-line",
  "supabase",
]

